Remove deprecated APIs and mark ZlibFilterStream as internal only
