+++ src/ZlibFilterStream.php
@@ -3,7 +3,7 @@
 namespace Clue\React\Zlib;
 
 /**
- * Compressor and decompressor using PHP's zlib compression filters.
+ * [Internal] Compressor and decompressor using PHP's zlib compression filters.
  *
  * Supports the following compression formats:
  *
@@ -11,67 +11,13 @@ namespace Clue\React\Zlib;
  * RFC 1951 (raw DEFLATE compressed format)
  * RFC 1950 (ZLIB compressed format)
  *
+ * @internal Should not be relied upon outside of this package.
  * @link http://php.net/manual/en/filters.compression.php
- * @deprecated 0.2.2 External usage of `ZlibFilterStream` is deprecated, use `Compressor` or `Decompressor` instead.
  * @see Compressor
  * @see Decompressor
  */
 class ZlibFilterStream extends TransformStream
 {
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createGzipCompressor($level = -1)
-    {
-        return new Compressor(ZLIB_ENCODING_GZIP, $level);
-    }
-
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createGzipDecompressor()
-    {
-        return new Decompressor(ZLIB_ENCODING_GZIP);
-    }
-
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createDeflateCompressor($level = -1)
-    {
-        return new Compressor(ZLIB_ENCODING_RAW, $level);
-    }
-
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createDeflateDecompressor()
-    {
-        return new Decompressor(ZLIB_ENCODING_RAW);
-    }
-
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createZlibCompressor($level = -1)
-    {
-        return new Compressor(ZLIB_ENCODING_DEFLATE, $level);
-    }
-
-    /**
-     * @deprecated
-     * @return self
-     */
-    public static function createZlibDecompressor()
-    {
-        return new Decompressor(ZLIB_ENCODING_DEFLATE);
-    }
-
     private $filter;
 
     /**
